requirements.txt
streamlit
google-genai
st-gsheets-connection
pandas